+++ cmsplugin_blocks/admin/__init__.py
@@ -1,0 +1,6 @@
+from .feature import FeatureAdmin
+
+
+__all__ = [
+    "FeatureAdmin",
+]