+++ cmsplugin_blocks/factories/__init__.py
@@ -1,6 +1,7 @@
 from .album import AlbumFactory, AlbumItemFactory
 from .card import CardFactory
 from .container import ContainerFactory
+from .feature import FeatureFactory
 from .hero import HeroFactory
 from .slider import SliderFactory, SlideItemFactory
 from .user import UserFactory
@@ -11,6 +12,7 @@ __all__ = [
     "AlbumItemFactory",
     "CardFactory",
     "ContainerFactory",
+    "FeatureFactory",
     "HeroFactory",
     "SliderFactory",
     "SlideItemFactory",