+++ cmsplugin_blocks/forms/__init__.py
@@ -1,6 +1,7 @@
 from .album import AlbumForm, AlbumItemForm
 from .card import CardForm
 from .container import ContainerForm
+from .feature import FeatureForm
 from .hero import HeroForm
 from .slider import SliderForm, SlideItemForm
 
@@ -10,6 +11,7 @@ __all__ = [
     "AlbumItemForm",
     "CardForm",
     "ContainerForm",
+    "FeatureForm",
     "HeroForm",
     "SliderForm",
     "SlideItemForm",