+++ cmsplugin_blocks/forms/card.py
@@ -1,21 +1,21 @@
 from django import forms
-from django.contrib.admin.widgets import FilteredSelectMultiple
+# from django.contrib.admin.widgets import FilteredSelectMultiple
 
 from djangocms_text_ckeditor.widgets import TextEditorWidget
 
-from ..choices_helpers import get_card_feature_choices
+# from ..choices_helpers import get_card_feature_choices
 from ..models.card import Card
 
 
 class CardForm(forms.ModelForm):
     # Enforce the right field since ModelAdmin ignore the formfield defined in custom
     # modelfield. Option have to fit to the modelfield ones.
-    features = forms.MultipleChoiceField(
-        choices=get_card_feature_choices(),
-        required=False,
-        widget=FilteredSelectMultiple(verbose_name=None, is_stacked=False),
-        # widget=forms.CheckboxSelectMultiple,
-    )
+    # size_features = forms.MultipleChoiceField(
+    #     choices=get_card_feature_choices(),
+    #     required=False,
+    #     widget=FilteredSelectMultiple(verbose_name=None, is_stacked=False),
+    #     widget=forms.CheckboxSelectMultiple,
+    # )
 
     class Meta:
         model = Card
@@ -23,7 +23,9 @@ class CardForm(forms.ModelForm):
         fields = [
             "title",
             "template",
-            "features",
+            "size_features",
+            "color_features",
+            "extra_features",
             "image",
             "image_alt",
             "content",
@@ -31,6 +33,9 @@ class CardForm(forms.ModelForm):
             "link_open_blank",
         ]
         widgets = {
+            "size_features": forms.CheckboxSelectMultiple,
+            "color_features": forms.CheckboxSelectMultiple,
+            "extra_features": forms.CheckboxSelectMultiple,
             "content": TextEditorWidget,
         }
 
@@ -38,11 +43,3 @@ class CardForm(forms.ModelForm):
         css = {
             "all": ("cmsplugin_blocks/css/admin/card.css",),
         }
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-        # Get back original model field name onto the field
-        self.fields["features"].label = (
-            self._meta.model._meta.get_field("features").verbose_name
-        )