+++ cmsplugin_blocks/models/__init__.py
@@ -1,6 +1,7 @@
 from .album import Album, AlbumItem
 from .card import Card
 from .container import Container
+from .feature import Feature
 from .hero import Hero
 from .slider import Slider, SlideItem
 
@@ -10,6 +11,7 @@ __all__ = [
     "AlbumItem",
     "Card",
     "Container",
+    "Feature",
     "Hero",
     "Slider",
     "SlideItem",