+++ cmsplugin_blocks/models/card.py
@@ -15,16 +15,11 @@ from smart_media.mixins import SmartFormatMixin
 from smart_media.modelfields import SmartMediaField
 from smart_media.signals import auto_purge_files_on_change
 
-from ..choices_helpers import (
-    get_card_feature_choices,
-    get_card_template_choices,
-    get_card_template_default,
-)
-from ..modelfields import CommaSeparatedStringsField
-from ..utils.validators import validate_css_classnames
+from ..choices_helpers import get_card_template_choices, get_card_template_default
+from .mixins import FeatureMixinModel
 
 
-class Card(SmartFormatMixin, CMSPlugin):
+class Card(SmartFormatMixin, FeatureMixinModel, CMSPlugin):
     """
     Card component.
     """
@@ -50,18 +45,6 @@ class Card(SmartFormatMixin, CMSPlugin):
     """
     Template choice from available plugin templates in setting
     ``BLOCKS_CARD_TEMPLATES``. Default to the first choice item.
-    """
-
-    features = CommaSeparatedStringsField(
-        _("Layout features"),
-        choices=get_card_feature_choices(),
-        blank=True,
-        default="",
-        max_length=255,
-        validators=[validate_css_classnames],
-    )
-    """
-    Optional string of CSS class names divided by a single comma.
     """
 
     image = SmartMediaField(
@@ -110,9 +93,43 @@ class Card(SmartFormatMixin, CMSPlugin):
         default=False,
         help_text=_("If checked the link will be opened in a new window"),
     )
+
     """
     Checkbox to enable opening link URL in a new window/tab.
     """
+    size_features = models.ManyToManyField(
+        "cmsplugin_blocks.Feature",
+        verbose_name=_("size features"),
+        related_name="%(app_label)s_%(class)s_size_related",
+        blank=True,
+        limit_choices_to={"scope": "size", "plugins__contains": "Card"},
+    )
+    """
+    Optional related size features.
+    """
+
+    color_features = models.ManyToManyField(
+        "cmsplugin_blocks.Feature",
+        verbose_name=_("color features"),
+        related_name="%(app_label)s_%(class)s_color_related",
+        blank=True,
+        limit_choices_to={"scope": "color", "plugins__contains": "Card"},
+    )
+    """
+    Optional related color features.
+    """
+
+    extra_features = models.ManyToManyField(
+        "cmsplugin_blocks.Feature",
+        verbose_name=_("extra features"),
+        related_name="%(app_label)s_%(class)s_extra_related",
+        blank=True,
+        limit_choices_to={"scope": "extra", "plugins__contains": "Card"},
+    )
+
+    class Meta:
+        verbose_name = _("Card")
+        verbose_name_plural = _("Cards")
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
@@ -126,22 +143,6 @@ class Card(SmartFormatMixin, CMSPlugin):
 
     def get_image_format(self):
         return self.media_format(self.image)
-
-    def get_features(self):
-        """
-        Merge feature items into a string with a comma divider.
-
-        Returns:
-            string: Feature items divided by a comma. Duplicate items are removed
-            and original order is preserved.
-        """
-        return " ".join(
-            list(dict.fromkeys(self.features))
-        )
-
-    class Meta:
-        verbose_name = _("Card")
-        verbose_name_plural = _("Cards")
 
 
 pre_save.connect(